cleanup broadcast feature
